pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-cli-assistant"
version = "2.0.0"
description = "Enhanced CLI AI assistant powered by Google Gen AI"
authors = [{name = "Patrick Larocque"}]
readme = "README.md"
requires-python = ">=3.14"
license = {text = "MIT"}
keywords = ["ai", "cli", "assistant", "gemini", "google-ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer[all]>=0.12.0",
    "pyyaml>=6.0.1",
    "tenacity>=8.2.0",
    "pygments>=2.17.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.scripts]
ai-assistant = "ai_cli_assistant.cli:main"

[project.urls]
Homepage = "https://github.com/patlar104/ai_cli_assistant"
Repository = "https://github.com/patlar104/ai_cli_assistant"
Issues = "https://github.com/patlar104/ai_cli_assistant/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["ai_cli_assistant*"]

[tool.setuptools.package-data]
"ai_cli_assistant" = ["prompts/*.txt", "prompts/examples/*.txt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=src/ai_cli_assistant --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py314"]

[tool.ruff]
line-length = 100
target-version = "py314"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
